Conflicts:
	.travis.yml

+++ src/arcemu-shared/Auth/WowCrypt.h
@@ -34,8 +34,8 @@ class WowCrypt
         ~WowCrypt();
 
         void Init(uint8* K);
-        ARCEMU_INLINE void DecryptRecv(uint8* pData, size_t len) { if(!m_initialized) { return; } RC4(&m_clientDecrypt, (unsigned long)len, pData, pData); }
-        ARCEMU_INLINE void EncryptSend(uint8* pData, size_t len) { if(!m_initialized) { return; } RC4(&m_serverEncrypt, (unsigned long)len, pData, pData); }
+        ARCEMU_INLINE void DecryptRecv(uint8* pData, size_t len) { if(!m_initialized) { return; } RC4(&m_clientDecrypt, len, pData, pData); }
+        ARCEMU_INLINE void EncryptSend(uint8* pData, size_t len) { if(!m_initialized) { return; } RC4(&m_serverEncrypt, len, pData, pData); }
         bool IsInitialized() { return m_initialized; }
 
     private:
