+++ src/arcemu-world/Corpse.cpp
@@ -88,7 +88,7 @@ void Corpse::DeleteFromDB()
     //delete corpse from db when its not needed anymore
     char sql[256];
 
-    snprintf(sql, 256, "DELETE FROM corpses WHERE guid=%u", (unsigned int)GetLowGUID());
+    snprintf(sql, 256, "DELETE FROM corpses WHERE guid=%u", GetLowGUID());
     CharacterDatabase.Execute(sql);
 }
 
@@ -96,7 +96,7 @@ void CorpseData::DeleteFromDB()
 {
     char sql[256];
 
-    snprintf(sql, 256, "DELETE FROM corpses WHERE guid=%u", (unsigned int)LowGuid);
+    snprintf(sql, 256, "DELETE FROM corpses WHERE guid=%u", LowGuid);
     CharacterDatabase.Execute(sql);
 }
