+++ src/arcemu-world/Pet.h
@@ -249,7 +249,7 @@ class SERVER_DECL Pet : public Creature
         AI_Spell* HandleAutoCastEvent();
         void SetPetSpellState(uint32 spell, uint16 state);
         void SetAutoCast(AI_Spell* sp, bool on);
-        float GetHappinessDmgMod() { return 0.25f * GetHappinessState() + 0.5f; };
+        float GetHappinessDmgMod() { return 0.25f * (float)GetHappinessState() + 0.5f; };
         bool IsBeingDeleted() { return ScheduledForDeletion; }
 
         virtual Group* GetGroup();
